requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aerich==0.7.1
aiosqlite==0.17.0; python_version >= '3.6'
anyio[trio]==3.6.2
async-generator==1.10; python_version >= '3.5'
async-timeout==4.0.2; python_version >= '3.6'
asyncmy==0.2.5
attrs==22.2.0; python_version >= '3.6'
bcrypt==4.0.1
certifi==2022.12.7; python_version >= '3.6'
cffi==1.15.1
click==8.1.3; python_version >= '3.7'
cryptography==39.0.0
dictdiffer==0.9.0
dnspython==2.2.1; python_version >= '3.6' and python_version < '4.0'
email-validator==1.3.0
exceptiongroup==1.1.0; python_version < '3.11'
fastapi[all]==0.89.1
gunicorn==20.1.0
h11==0.14.0; python_version >= '3.7'
hiredis==2.1.1; python_version >= '3.7'
httpcore==0.16.3; python_version >= '3.7'
httptools==0.5.0
httpx==0.23.3
idna==3.4; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
iso8601==1.1.0; python_version < '4.0' and python_full_version >= '3.6.2'
itsdangerous==2.1.2
jinja2==3.1.2
loguru==0.6.0
markupsafe==2.1.1; python_version >= '3.7'
mongoengine==0.25.0
mongomock==4.1.2
more-itertools==8.14.0; python_version >= '3.5'
orjson==3.8.5
outcome==1.2.0; python_version >= '3.7'
packaging==23.0; python_version >= '3.7'
phonenumbers==8.13.4
pluggy==1.0.0; python_version >= '3.6'
pptree==3.1
pycparser==2.21
pydantic==1.10.4; python_version >= '3.7'
pyjwt==2.6.0
pymongo==4.3.3; python_version >= '3.7'
pypika-tortoise==0.1.6; python_version >= '3.7' and python_version < '4.0'
pytest==7.2.0
python-dotenv==0.21.0
python-multipart==0.0.5
python-ulid==1.1.0; python_version >= '3.7'
pytz==2022.7
pyyaml==6.0
redis-om==0.1.2
redis==4.4.1; python_version >= '3.7'
rfc3986[idna2008]==1.5.0
sentinels==1.0.0
setuptools==65.6.3; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
sortedcontainers==2.4.0
starlette==0.22.0; python_version >= '3.7'
tomli==2.0.1; python_version < '3.11'
tomlkit==0.11.6; python_version >= '3.6'
tortoise-orm[asyncmy]==0.19.2
trio==0.21.0
types-redis==4.4.0.0
typing-extensions==4.4.0; python_version < '3.10'
ujson==5.7.0; python_version >= '3.7'
uvicorn[standard]==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
